+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/remove_service/remove_service_test.go
@@ -8,7 +8,10 @@ import (
 
 func TestRemoveService_StringRepresentation(t *testing.T) {
 	removeInstruction := NewRemoveServiceInstruction(nil, *kurtosis_instruction.NewInstructionPosition(4, 4, "dummyFile"), "dummy-service-id")
-	expectedStrRep := `remove_service(service_id="dummy-service-id")`
+	expectedStrRep := `// from: dummyFile-4:4
+remove_service(
+	service_id="dummy-service-id",
+)`
 	require.Equal(t, expectedStrRep, removeInstruction.String())
 	require.Equal(t, expectedStrRep, removeInstruction.GetCanonicalInstruction())
 }